Adapt to refactor
